Add get_ticketing_url_from_description function.
The function collects the URLs from the given string (HTML or plain text, trying both), removes those that are obviously not ticketing links (ex: videoconference) and if more than one is left, filters further to just those having works like "registration" or "ticket" in the anchor text.

This changelist adds a test suite as well, which for now only contains test cases for the new function. It's a convenient place to copy/paste descriptions on which the function fails without restarting a scrape of the universe.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>