feat: parse ticketing link from iCal event description (#59)
* Add parsing of iCalendar events using https://icspy.readthedocs.io/en/stable/index.html.

Events starting before the current date and from which no ticketing link can be extracted are skipped.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* Add workshops which publish to the iCalendar format.

These are Planet C, Fresque du Sol and Fresque des Déchets.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* Fresque des Dechets already has ID 15.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* Restore Fresque des Dechets with proper ID, declare ICS dependency to Nix.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* Add missing closing quote.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* Add get_ticketing_url_from_description function.

The function collects the URLs from the given string (HTML or plain text, trying both), removes those that are obviously not ticketing links (ex: videoconference) and if more than one is left, filters further to just those having works like "registration" or "ticket" in the anchor text.

This changelist adds a test suite as well, which for now only contains test cases for the new function. It's a convenient place to copy/paste descriptions on which the function fails without restarting a scrape of the universe.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

---------

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>