chore: EventBrite updates (#78)
* Update to new EventBrite tags, scrape new EventBrite date format, don't crash on addresses without a newline.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* chore: update date and location tags

The class starting with `Location-module__addressWrapper___` contains the full address rather than just the address "title" which is the address line.

* fix: add missing iframe for Fresque de l'Economie Circulaire.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* feat: parse dates in German, parse minutes instead of just the hour.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* feat: support one more offline event type.

EventBrite events have two locations, a "short" one at the top and a full one with address and a lot more text like routing directions at the bottom. The elements and therefore the CSS selectors
vary depending on the event being offline (just text) or online (link to another element on the page).

The scraper now detects whether an element is online by searching for the short location (CSS selector `span.start-date-and-location__location`) which is present in both cases.

The scraper also doesn't crash if it expects an offline element and cannot find the full location. Instead, it logs an error and continues to the next event. This can happen when the scraper
doesn't detect that the event is online.

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

* fix: don't add 12 hours to 12 PM, parse German UTC (MEZ)

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>

---------

Signed-off-by: Jeffrey Belt <jeffrey@theshifters.ch>